Restructured folders, files, and removed Test file
