+++ frontend/components/home/InfoPanel.js
