--- frontend/components/styles/Test.js
@@ -1,7 +1,0 @@
-import styled from 'styled-components';
-
-const Test = styled.h3`
-  color: red;
-`;
-
-export default Test;
